refactor(types): migrate from tsd to tstyche
- rename type test file to .tst.ts and update assertions to tstyche API
- remove the dedicated tsd config and update package scripts/dependencies
- move type definitions from root to types/ folder
